fix: prevent app route always changes to root after refresh
