+++ src/__tests__/App.test.tsx
@@ -1,9 +1,19 @@
 import { shallow } from 'enzyme'
 import React from 'react'
+import { Provider } from 'react-redux'
+import configureStore from 'redux-mock-store'
 
 import App from '../App'
 
 it('renders without crashing', () => {
-  const wrapper = shallow(<App />)
+  const mockStore = configureStore()({})
+
+  const AppWithStore = () => (
+    <Provider store={mockStore}>
+      <App />
+    </Provider>
+  )
+
+  const wrapper = shallow(<AppWithStore />)
   expect(wrapper).toBeDefined()
 })
