+++ src/index.tsx
@@ -1,14 +1,20 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-
+import { Provider } from 'react-redux'
 import '@hospitalrun/components/scss/main.scss'
+
 import './index.css'
 import App from './App'
 import * as serviceWorker from './serviceWorker'
-
 import './shared/config/i18n'
+import store from './shared/store'
 
-ReactDOM.render(<App />, document.getElementById('root'))
+ReactDOM.render(
+  <Provider store={store}>
+    <App />
+  </Provider>,
+  document.getElementById('root'),
+)
 
 // If you want your app to work offline and load faster, you can change
 // unregister() to register() below. Note this comes with some pitfalls.
