+++ src/login/Login.tsx
@@ -1,37 +1,19 @@
 import { Alert, Container, Panel } from '@hospitalrun/components'
-import React, { useEffect, useState } from 'react'
+import React, { useState } from 'react'
 import Button from 'react-bootstrap/Button'
 import { useDispatch, useSelector } from 'react-redux'
 import { Redirect } from 'react-router-dom'
 
 import TextInputWithLabelFormGroup from '../shared/components/input/TextInputWithLabelFormGroup'
-import { remoteDb } from '../shared/config/pouchdb'
 import logo from '../shared/static/images/logo-on-transparent.png'
 import { RootState } from '../shared/store'
-import { getCurrentSession, login } from '../user/user-slice'
+import { login } from '../user/user-slice'
 
 const Login = () => {
   const dispatch = useDispatch()
   const [username, setUsername] = useState('')
   const [password, setPassword] = useState('')
   const { loginError, user } = useSelector((root: RootState) => root.user)
-  const [loading, setLoading] = useState(true)
-
-  useEffect(() => {
-    const init = async () => {
-      try {
-        const session = await remoteDb.getSession()
-        if (session.userCtx.name) {
-          await dispatch(getCurrentSession(session.userCtx.name))
-        }
-      } catch (e) {
-        console.log(e)
-      }
-      setLoading(false)
-    }
-
-    init()
-  }, [dispatch])
 
   const onUsernameChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     const { value } = event.currentTarget
@@ -47,43 +29,37 @@ const Login = () => {
     await dispatch(login(username, password))
   }
 
-  if (loading) {
-    return null
-  }
-
   if (user) {
     return <Redirect to="/" />
   }
 
   return (
-    <>
-      <Container className="container align-items-center" style={{ width: '50%' }}>
-        <img src={logo} alt="HospitalRun" style={{ width: '100%', textAlign: 'center' }} />
-        <form>
-          <Panel title="Please Sign In" color="primary">
-            {loginError && <Alert color="danger" message={loginError} title="Unable to login" />}
-            <TextInputWithLabelFormGroup
-              isEditable
-              label="username"
-              name="username"
-              value={username}
-              onChange={onUsernameChange}
-            />
-            <TextInputWithLabelFormGroup
-              isEditable
-              type="password"
-              label="password"
-              name="password"
-              value={password}
-              onChange={onPasswordChange}
-            />
-            <Button block onClick={onSignInClick}>
-              Sign In
-            </Button>
-          </Panel>
-        </form>
-      </Container>
-    </>
+    <Container className="container align-items-center" style={{ width: '50%' }}>
+      <img src={logo} alt="HospitalRun" style={{ width: '100%', textAlign: 'center' }} />
+      <form>
+        <Panel title="Please Sign In" color="primary">
+          {loginError && <Alert color="danger" message={loginError} title="Unable to login" />}
+          <TextInputWithLabelFormGroup
+            isEditable
+            label="username"
+            name="username"
+            value={username}
+            onChange={onUsernameChange}
+          />
+          <TextInputWithLabelFormGroup
+            isEditable
+            type="password"
+            label="password"
+            name="password"
+            value={password}
+            onChange={onPasswordChange}
+          />
+          <Button block onClick={onSignInClick}>
+            Sign In
+          </Button>
+        </Panel>
+      </form>
+    </Container>
   )
 }
